Fixed some integer precision stuff and did a minor fix to note handling
+ Loss of integer precision in note ID parsing has been fixed
+ Notifications for incoming notes would incorrectly say that they were sent to whichever character is logged into F-Chat. This isn’t necessarily true, so the notification has been fixed accordingly.

+++ ESFlistNoteTrampoline.h
@@ -7,7 +7,7 @@
 #ifndef F_Chat_ESFlistNoteTrampoline_h
 #define F_Chat_ESFlistNoteTrampoline_h
 
-void SendFlistNote(char *message, char *url, PurpleConnection *pc, int noteID);
+void SendFlistNote(char *message, char *url, PurpleConnection *pc, long noteID);
 
 
 #endif
